requirements.txt
flask>=2.3.0
spotipy>=2.23.0
pillow>=10.0.0
qrcode[pil]>=7.4.2
requests>=2.31.0
python-dotenv>=1.0.0
